Move raise_fd_limit code to subspace-networking
