+++ crates/subspace-gateway/src/main.rs
@@ -7,9 +7,10 @@ mod node_client;
 mod piece_getter;
 mod piece_validator;
 
-use crate::commands::{Command, raise_fd_limit, set_exit_on_panic};
+use crate::commands::{Command, set_exit_on_panic};
 use clap::Parser;
 use subspace_logging::init_logger;
+use subspace_networking::utils::raise_fd_limit;
 use tracing::info;
 
 #[global_allocator]
